update our flakes.

+++ pkg/hwinfo/detail_monitor.go
@@ -7,30 +7,30 @@ package hwinfo
 import "C"
 
 type SyncRange struct {
-	Min uint `json:"min"`
-	Max uint `json:"max"`
+	Min uint16 `json:"min"`
+	Max uint16 `json:"max"`
 }
 
 type SyncTimings struct {
-	Disp      uint `json:"disp"` // todo what's the proper name for this?
-	SyncStart uint `json:"sync_start"`
-	SyncEnd   uint `json:"sync_end"`
-	Total     uint `json:"total"` // todo what's a better name for this?
+	Disp      uint16 `json:"disp"` // todo what's the proper name for this?
+	SyncStart uint16 `json:"sync_start"`
+	SyncEnd   uint16 `json:"sync_end"`
+	Total     uint32 `json:"total"` // todo what's a better name for this?
 }
 
 type DetailMonitor struct {
 	Type                  DetailType  `json:"-"`
-	ManufactureYear       uint        `json:"manufacture_year"`
-	ManufactureWeek       uint        `json:"manufacture_week"`
+	ManufactureYear       uint16      `json:"manufacture_year"`
+	ManufactureWeek       uint8       `json:"manufacture_week"`
 	VerticalSync          SyncRange   `json:"vertical_sync"`
 	HorizontalSync        SyncRange   `json:"horizontal_sync"`
 	HorizontalSyncTimings SyncTimings `json:"horizontal_sync_timings"`
 	VerticalSyncTimings   SyncTimings `json:"vertical_sync_timings"`
-	Clock                 uint        `json:"clock"`
-	Width                 uint        `json:"width"`
-	Height                uint        `json:"height"`
-	WidthMillimetres      uint        `json:"width_millimetres"`
-	HeightMillimetres     uint        `json:"height_millimetres"`
+	Clock                 uint32      `json:"clock"`
+	Width                 uint32      `json:"width"`
+	Height                uint32      `json:"height"`
+	WidthMillimetres      uint32      `json:"width_millimetres"`
+	HeightMillimetres     uint32      `json:"height_millimetres"`
 	HorizontalFlag        byte        `json:"horizontal_flag"`
 	VerticalFlag          byte        `json:"vertical_flag"`
 	Vendor                string      `json:"vendor"`
@@ -48,32 +48,32 @@ func NewDetailMonitor(mon C.hd_detail_monitor_t) (*DetailMonitor, error) {
 
 	return &DetailMonitor{
 		Type:            DetailTypeMonitor,
-		ManufactureYear: uint(data.manu_year),
-		ManufactureWeek: uint(data.manu_week),
+		ManufactureYear: uint16(data.manu_year),
+		ManufactureWeek: uint8(data.manu_week),
 		VerticalSync: SyncRange{
-			Min: uint(data.min_vsync),
-			Max: uint(data.max_vsync),
+			Min: uint16(data.min_vsync),
+			Max: uint16(data.max_vsync),
 		},
 		HorizontalSync: SyncRange{
-			Min: uint(data.min_hsync),
-			Max: uint(data.max_hsync),
+			Min: uint16(data.min_hsync),
+			Max: uint16(data.max_hsync),
 		},
-		Clock:             uint(data.clock),
-		Width:             uint(data.width),
-		Height:            uint(data.height),
-		WidthMillimetres:  uint(data.width_mm),
-		HeightMillimetres: uint(data.height_mm),
+		Clock:             uint32(data.clock),
+		Width:             uint32(data.width),
+		Height:            uint32(data.height),
+		WidthMillimetres:  uint32(data.width_mm),
+		HeightMillimetres: uint32(data.height_mm),
 		HorizontalSyncTimings: SyncTimings{
-			Disp:      uint(data.hdisp),
-			SyncStart: uint(data.hsyncstart),
-			SyncEnd:   uint(data.hsyncend),
-			Total:     uint(data.htotal),
+			Disp:      uint16(data.hdisp),
+			SyncStart: uint16(data.hsyncstart),
+			SyncEnd:   uint16(data.hsyncend),
+			Total:     uint32(data.htotal),
 		},
 		VerticalSyncTimings: SyncTimings{
-			Disp:      uint(data.vdisp),
-			SyncStart: uint(data.vsyncstart),
-			SyncEnd:   uint(data.vsyncend),
-			Total:     uint(data.vtotal),
+			Disp:      uint16(data.vdisp),
+			SyncStart: uint16(data.vsyncstart),
+			SyncEnd:   uint16(data.vsyncend),
+			Total:     uint32(data.vtotal),
 		},
 		HorizontalFlag: byte(data.hflag),
 		VerticalFlag:   byte(data.vflag),
